auth-backend: move a couple more types to auth-node
